requirements.txt
crewai
crewai_tools
langchain
python-dotenv
langchain_groq
langchain_community
agentops